Cleanup index.js and push members down to crux-ui package
